refactor(core): split types.ts into domain modules behind a barrel
types.ts had grown to 730 lines mixing eight unrelated domains. It is now
src/core/types/ with one module per domain - layout (grid data model +
Coord/Rect), baseplate (StackPrintParams, StoredBaseplateParams),
drawerOutline, interaction (handles/swap/Interaction), validation
(placement reasons + BlockedZone), printList, preview (ThumbnailBin/
LayoutPreview), snapshots, share, library - re-exported by types/index.ts
together with the branded-type re-exports, so every existing
'@/core/types' import (551 files) resolves unchanged.

Two latent type-only cycles fell out of the grouping and are broken
structurally: PaddingAnchor moved into baseplate (layout <-> baseplate)
and the preview types got their own module (share <-> library).

Code moved verbatim; the runtime members (DEFAULT_MAGNET_ANCHOR,
effectiveGridUnitMmY, STACK_PRINT_*) keep their import paths via the
barrel.
